:recycle: refactor : mapper 네이밍을 Response에서 WebAdapter Mapper로 수정
