+++ src/main/kotlin/org/example/codereviewstudy/infrastructure/web/rest/post/mapper/PostWebAdapterMapper.kt
@@ -1,9 +1,13 @@
 package org.example.codereviewstudy.infrastructure.web.rest.post.mapper
 
 import org.example.codereviewstudy.domain.post.model.Post
+import org.example.codereviewstudy.domain.user.model.User
+import org.example.codereviewstudy.infrastructure.web.rest.post.request.PostCreateRequest
 import org.example.codereviewstudy.infrastructure.web.rest.post.response.PostCreateResponse
 import org.example.codereviewstudy.infrastructure.web.rest.post.response.PostResponse
 import org.example.codereviewstudy.infrastructure.web.rest.post.response.PostUpdateResponse
+
+import java.time.LocalDateTime
 
 fun Post.toCreateResponse(): PostCreateResponse {
     return PostCreateResponse(
@@ -32,5 +36,14 @@ fun Post.toUpdateResponse(): PostUpdateResponse {
         content = content,
         createdAt = createdAt,
         id = id,
+    )
+}
+
+fun PostCreateRequest.toDomain(author: User): Post {
+    return Post(
+        title = title,
+        content = content,
+        author = author,
+        createdAt = LocalDateTime.now(),
     )
 }
